+++ example/example.dart
@@ -3,10 +3,10 @@
 
 import 'package:cli_repl/cli_repl.dart';
 
-main(args) {
+main(args) async {
   var v = (str) => str.trim().isEmpty || str.trim().endsWith(';');
   var repl = new Repl(prompt: '>>> ', continuation: '... ', validator: v);
-  for (var x in repl.run()) {
+  await for (var x in repl.runAsync()) {
     if (x.trim().isEmpty) continue;
     print(x);
   }